+++ doc/changelog.md
@@ -1,5 +1,10 @@
 # Changes
 
+## v1.2.1 (UNRELEASED)
+
+- New: Add an accessor for the configured clean block
+- New: Add a hook to use fabricated durations instead of actual timing data.
+
 ## v1.2.0 (5 July 2018)
 
 - New: Use monotonic clock for duration calculations